Solve clippy warnings

+++ chains/container-chains/runtime-templates/simple/src/lib.rs
@@ -711,10 +711,10 @@ where
     ) -> Option<UncheckedExtrinsic> {
         use sp_runtime::traits::StaticLookup;
         // take the biggest period possible.
-        let period = BlockHashCount::get()
+        let period = u64::from(BlockHashCount::get()
             .checked_next_power_of_two()
             .map(|c| c / 2)
-            .unwrap_or(2) as u64;
+            .unwrap_or(2));
 
         let current_block = System::block_number()
             .saturated_into::<u64>()
@@ -1174,6 +1174,8 @@ impl_runtime_apis! {
                             initial_asset_amount - asset_amount,
                         );
                     });
+                    
+                    #[allow(clippy::cast_sign_loss)]
                     Some((assets, fee_index as u32, dest, verify))
                 }
             }
